tests/wpt/tests/html/canvas/offscreen/canvas-host/2d.canvas.host.initial.reset.same.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<title>OffscreenCanvas test: 2d.canvas.host.initial.reset.same</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.canvas.host.initial.reset.same</h1>
<p class="desc">Setting size (not changing the value) resets canvas to transparent black</p>


<script>
var t = async_test("Setting size (not changing the value) resets canvas to transparent black");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

  var canvas = new OffscreenCanvas(100, 50);
  var ctx = canvas.getContext('2d');

  canvas.width = 100;
  ctx.fillStyle = '#f00';
  ctx.fillRect(0, 0, 50, 50);
  _assertPixel(canvas, 20,20, 255,0,0,255);
  canvas.width = 100;
  _assertPixel(canvas, 20,20, 0,0,0,0);
  t.done();

});
</script>
